fix : 클래스명 구체적 역할이 알 수 있도록 변경
