+++ src/main/java/com/flab/goodchoice/coupon/api/CouponController.java
@@ -1,6 +1,7 @@
 package com.flab.goodchoice.coupon.api;
 
-import com.flab.goodchoice.coupon.application.CouponService;
+import com.flab.goodchoice.coupon.application.CouponCommandService;
+import com.flab.goodchoice.coupon.application.CouponQueryService;
 import com.flab.goodchoice.coupon.dto.CouponInfoResponse;
 import com.flab.goodchoice.coupon.dto.CreateCouponRequest;
 import com.flab.goodchoice.coupon.dto.ModifyCouponRequest;
@@ -14,34 +15,41 @@ import java.util.UUID;
 @RestController
 public class CouponController {
 
-    private final CouponService couponService;
+    private final CouponQueryService couponQueryService;
+    private final CouponCommandService couponCommandService;
 
-    public CouponController(final CouponService couponService) {
-        this.couponService = couponService;
+    public CouponController(CouponQueryService couponQueryService, CouponCommandService couponCommandService) {
+        this.couponQueryService = couponQueryService;
+        this.couponCommandService = couponCommandService;
     }
 
     @PostMapping
     public UUID create(@RequestBody final CreateCouponRequest createCouponRequest) {
         validation(createCouponRequest.couponName(), createCouponRequest.stock());
 
-        return couponService.create(createCouponRequest.couponName(), createCouponRequest.stock());
+        return couponCommandService.create(createCouponRequest.couponName(), createCouponRequest.stock());
     }
 
     @GetMapping
     public List<CouponInfoResponse> getAllCoupons() {
-        return couponService.getAllCoupons();
+        return couponQueryService.getAllCoupons();
     }
 
     @GetMapping("/{couponToken}")
     public CouponInfoResponse getCouponDetail(@PathVariable final UUID couponToken) {
-        return couponService.getCouponDetail(couponToken);
+        return couponQueryService.getCouponDetail(couponToken);
     }
 
     @PutMapping("/{couponToken}")
     public UUID modifyCoupon(@PathVariable final UUID couponToken, @RequestBody final ModifyCouponRequest modifyCouponRequest) {
         validation(modifyCouponRequest.couponName(), modifyCouponRequest.stock());
 
-        return couponService.modifyCoupon(couponToken, modifyCouponRequest.couponName(), modifyCouponRequest.stock());
+        return couponCommandService.modifyCoupon(couponToken, modifyCouponRequest.couponName(), modifyCouponRequest.stock());
+    }
+
+    @DeleteMapping("/{couponToken}")
+    public UUID deleteCoupon(@PathVariable final UUID couponToken) {
+        return couponCommandService.deleteCoupon(couponToken);
     }
 
     private void validation(final String couponName, final int stock) {
@@ -52,10 +60,5 @@ public class CouponController {
         if (stock < 0) {
             throw new IllegalArgumentException("쿠폰 갯수는 음수가 될수 없습니다.");
         }
-    }
-
-    @DeleteMapping("/{couponToken}")
-    public UUID deleteCoupon(@PathVariable final UUID couponToken) {
-        return couponService.deleteCoupon(couponToken);
     }
 }
